carrito a la mitad

+++ routes/web.php
@@ -5,21 +5,23 @@ Route::get('/', 'TestController@welcome');
 Auth::routes();
 
 Route::get('/home', 'HomeController@index')->name('home');
+Route::get('products/{id}', 'ProductController@show');
 
+Route::post('/cart', 'CartDetailController@store');
 Route::middleware(['auth','admin'])->group(function () {
-    Route::get('/admin/products', 'ProductController@index'); //listar
-	Route::get('/admin/products/create', 'ProductController@create'); //abrir form crear
-	Route::post('/admin/products', 'ProductController@store'); //insertar
+    Route::get('/admin/products', 'Admin\ProductController@index'); //listar
+	Route::get('/admin/products/create', 'Admin\ProductController@create'); //abrir form crear
+	Route::post('/admin/products', 'Admin\ProductController@store'); //insertar
 
-	Route::get('/admin/products/{id}/edit', 'ProductController@edit'); //abrir form editar
-	Route::post('/admin/products/{id}/edit', 'ProductController@update');//editar
+	Route::get('/admin/products/{id}/edit', 'Admin\ProductController@edit'); //abrir form editar
+	Route::post('/admin/products/{id}/edit', 'Admin\ProductController@update');//editar
 
-	Route::post('/admin/products/{id}/delete', 'ProductController@destroy'); //eliminar
+	Route::post('/admin/products/{id}/delete', 'Admin\ProductController@destroy'); //eliminar
 
 	//imagenes
-	Route::get('/admin/products/{id}/images', 'ImageController@index');//lista
-	Route::post('/admin/products/{id}/images', 'ImageController@store'); //insertar
-	Route::delete('/admin/products/{id}/images', 'ImageController@destroy'); //borrar
-	Route::get('/admin/products/{id}/images/select/{image}', 'ImageController@select'); //destacar
+	Route::get('/admin/products/{id}/images', 'Admin\ImageController@index');//lista
+	Route::post('/admin/products/{id}/images', 'Admin\ImageController@store'); //insertar
+	Route::delete('/admin/products/{id}/images', 'Admin\ImageController@destroy'); //borrar
+	Route::get('/admin/products/{id}/images/select/{image}', 'Admin\ImageController@select'); //destacar
 
 });
